SlutaSnusaFrontend/requirements.txt
Django
requests
datetime
python-dateutil